refactor: move secondary ports to a dedicated directory
