+++ __tests__/unit/template.provider.spec.ts
@@ -1,4 +1,4 @@
-import { TemplateProvider } from 'src/domain/ports/templates.provider.js';
+import { TemplateProvider } from 'src/domain/ports/secondary/templates.provider.js';
 import { JsxTemplateProvider } from 'src/secondary/adapters/template.provider.js';
 
 describe('JsxTemplateProvider (unit)', () => {
