+++ src/domain/ports/secondary/weather.provider.ts
@@ -1,4 +1,4 @@
-import { Weather } from '../entities/weather.entity.js';
+import { Weather } from 'src/domain/entities/weather.entity.js';
 
 export interface WeatherProvider {
   getWeather(city: string): Promise<Weather>;
