+++ src/primary/adapters/app.ts
@@ -3,8 +3,8 @@ import { ServerType } from '@hono/node-server';
 import { Server } from './server.js';
 import { promisify } from 'node:util';
 import { DbProvider } from 'src/secondary/adapters/db.provider.js';
-import { Logger, LoggerProvider } from 'src/domain/ports/logger.provider.js';
-import { CronProvider } from 'src/domain/ports/cron.provider.js';
+import { Logger, LoggerProvider } from 'src/domain/ports/secondary/logger.provider.js';
+import { CronProvider } from 'src/domain/ports/secondary/cron.provider.js';
 
 const GRACEFUL_SHUTDOWN_DELAY = 15_000;
