+++ src/secondary/adapters/subscription.repository.ts
@@ -2,7 +2,7 @@ import { eq } from 'drizzle-orm';
 import { subscriptions } from 'src/secondary/db.schema.js';
 import { Exception } from 'src/domain/entities/exception.entity.js';
 import { Db, DbProvider } from './db.provider.js';
-import { SubscriptionRepository } from 'src/domain/ports/subscription.repository.js';
+import { SubscriptionRepository } from 'src/domain/ports/secondary/subscription.repository.js';
 import { CreateSubscriptionDto } from 'src/domain/dtos/create-subscription.dto.js';
 import { Frequency } from 'src/domain/entities/subscription.entity.js';
