+++ src/secondary/adapters/weather/cache.provider.ts
@@ -1,5 +1,5 @@
 import { CacheProvider } from '../cache.provider.js';
-import { WeatherProvider } from 'src/domain/ports/weather.provider.js';
+import { WeatherProvider } from 'src/domain/ports/secondary/weather.provider.js';
 import { Weather } from 'src/domain/entities/weather.entity.js';
 import { Exception, ExceptionCode } from 'src/domain/entities/exception.entity.js';
