+++ src/secondary/adapters/weather/open-meteo.provider.ts
@@ -1,5 +1,5 @@
 import { Exception, ExceptionCode } from 'src/domain/entities/exception.entity.js';
-import { WeatherProvider } from 'src/domain/ports/weather.provider.js';
+import { WeatherProvider } from 'src/domain/ports/secondary/weather.provider.js';
 import { Weather } from 'src/domain/entities/weather.entity.js';
 import { HttpProvider } from '../http/http.provider.js';
